+++ src/client/java/com/github/theredbrain/rpgcrafting/gui/screen/ingame/HandCraftingScreen.java
@@ -147,9 +147,9 @@ public class HandCraftingScreen extends HandledScreen<HandCraftingScreenHandler>
 
 			this.handler.getCraftingResultInventory().addStack(craftingRecipeEntry.value().result.copy());
 
-			List<ItemStack> itemStackIngredients = craftingRecipeEntry.value().itemStackIngredients;
-			for (ItemStack itemStackIngredient : itemStackIngredients) {
-				this.handler.getCraftingResultIngredientsInventory().addStack(itemStackIngredient.copy());
+			List<RPGCraftingRecipe.ItemStackIngredient> itemStackIngredients = craftingRecipeEntry.value().itemStackIngredients;
+			for (RPGCraftingRecipe.ItemStackIngredient itemStackIngredient : itemStackIngredients) {
+				this.handler.getCraftingResultIngredientsInventory().addStack(itemStackIngredient.itemStack().copy());
 			}
 
 			List<Ingredient> ingredients = craftingRecipeEntry.value().ingredients;