+++ src/main/java/com/github/theredbrain/rpgcrafting/recipe/RPGCraftingRecipe.java
@@ -5,11 +5,16 @@ import com.github.theredbrain.rpgcrafting.recipe.input.MultipleStackRecipeInput;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
+import io.netty.handler.codec.DecoderException;
+import io.netty.handler.codec.EncoderException;
+import net.minecraft.component.ComponentChanges;
+import net.minecraft.component.ComponentType;
 import net.minecraft.inventory.Inventory;
 import net.minecraft.inventory.SimpleInventory;
 import net.minecraft.item.ItemStack;
 import net.minecraft.network.RegistryByteBuf;
 import net.minecraft.network.codec.PacketCodec;
+import net.minecraft.network.codec.PacketCodecs;
 import net.minecraft.recipe.Ingredient;
 import net.minecraft.recipe.Recipe;
 import net.minecraft.recipe.RecipeSerializer;
@@ -20,9 +25,12 @@ import net.minecraft.world.World;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
 
 public class RPGCraftingRecipe implements Recipe<MultipleStackRecipeInput> {
-	public final List<ItemStack> itemStackIngredients;
+	public final List<ItemStackIngredient> itemStackIngredients;
 	public final List<Ingredient> ingredients;
 	public final ItemStack result;
 	public int level;
@@ -31,7 +39,7 @@ public class RPGCraftingRecipe implements Recipe<MultipleStackRecipeInput> {
 	public final boolean showNotification;
 	public final boolean requiresUnlockAdvancement;
 
-	public RPGCraftingRecipe(List<ItemStack> itemStackIngredients, List<Ingredient> ingredients, ItemStack result, int level, int tab, String recipeType, boolean showNotification, boolean requiresUnlockAdvancement) {
+	public RPGCraftingRecipe(List<ItemStackIngredient> itemStackIngredients, List<Ingredient> ingredients, ItemStack result, int level, int tab, String recipeType, boolean showNotification, boolean requiresUnlockAdvancement) {
 		this.itemStackIngredients = itemStackIngredients;
 		this.ingredients = ingredients;
 		this.result = result;
@@ -54,11 +62,11 @@ public class RPGCraftingRecipe implements Recipe<MultipleStackRecipeInput> {
 			playerInventoryCopy.setStack(j, input.getStackInSlot(j).copy());
 		}
 
-		for (ItemStack itemStackIngredient : this.itemStackIngredients) {
-			int ingredientCount = itemStackIngredient.getCount();
+		for (ItemStackIngredient itemStackIngredient : this.itemStackIngredients) {
+			int ingredientCount = itemStackIngredient.itemStack.getCount();
 			for (j = 0; j < inputSize; j++) {
 				bl = false;
-				if (areItemStacksEqual(itemStackIngredient, playerInventoryCopy.getStack(j))) {
+				if (checkItemStackIngredient(itemStackIngredient, playerInventoryCopy.getStack(j))) {
 					itemStack = playerInventoryCopy.getStack(j).copy();
 					int stackCount = itemStack.getCount();
 					if (stackCount > ingredientCount) {
@@ -112,9 +120,9 @@ public class RPGCraftingRecipe implements Recipe<MultipleStackRecipeInput> {
 			inputCopy.setStack(j, input.getStackInSlot(j).copy());
 		}
 
-		for (ItemStack itemStackIngredient : this.itemStackIngredients) {
+		for (ItemStackIngredient itemStackIngredient : this.itemStackIngredients) {
 			for (j = 0; j < inputSize; j++) {
-				if (areItemStacksEqual(itemStackIngredient, inputCopy.getStack(j))) {
+				if (checkItemStackIngredient(itemStackIngredient, inputCopy.getStack(j))) {
 					return true;
 				}
 			}
@@ -129,16 +137,16 @@ public class RPGCraftingRecipe implements Recipe<MultipleStackRecipeInput> {
 		return false;
 	}
 
-	public static boolean areItemStacksEqual(ItemStack itemStack1, ItemStack itemStack2) {
-		if (itemStack1.getComponentChanges().isEmpty()) {
-			return itemStack1.getItem() == itemStack2.getItem();
-		} else {
-			if (itemStack1.getItem() == itemStack2.getItem()) {
-				return ItemStack.areItemsAndComponentsEqual(itemStack1, itemStack2);
-			}
-		}
-		return false;
-	}
+//	public static boolean areItemStacksEqual(ItemStack itemStack1, ItemStack itemStack2) {
+//		if (itemStack1.getComponentChanges().isEmpty()) {
+//			return itemStack1.getItem() == itemStack2.getItem();
+//		} else {
+//			if (itemStack1.getItem() == itemStack2.getItem()) {
+//				return ItemStack.areItemsAndComponentsEqual(itemStack1, itemStack2);
+//			}
+//		}
+//		return false;
+//	}
 
 	@Override
 	public ItemStack craft(MultipleStackRecipeInput input, RegistryWrapper.WrapperLookup lookup) {
@@ -181,7 +189,7 @@ public class RPGCraftingRecipe implements Recipe<MultipleStackRecipeInput> {
 
 		public static final MapCodec<RPGCraftingRecipe> CODEC = RecordCodecBuilder.mapCodec(
 				instance -> instance.group(
-						ItemStack.VALIDATED_CODEC.listOf().fieldOf("itemStackIngredients").forGetter(recipe -> recipe.itemStackIngredients),
+						ItemStackIngredient.CODEC.listOf().fieldOf("itemStackIngredients").forGetter(recipe -> recipe.itemStackIngredients),
 						Ingredient.DISALLOW_EMPTY_CODEC.listOf().fieldOf("ingredients").forGetter(recipe -> recipe.ingredients),
 						ItemStack.VALIDATED_CODEC.fieldOf("result").forGetter(recipe -> recipe.result),
 						Codec.INT.optionalFieldOf("level", 0).forGetter(recipe -> recipe.level),
@@ -207,9 +215,9 @@ public class RPGCraftingRecipe implements Recipe<MultipleStackRecipeInput> {
 
 		private static RPGCraftingRecipe read(RegistryByteBuf buf) {
 			int itemStackIngredientsSize = buf.readInt();
-			List<ItemStack> itemStackIngredients = new ArrayList<>();
+			List<ItemStackIngredient> itemStackIngredients = new ArrayList<>();
 			for (int i = 0; i < itemStackIngredientsSize; i++) {
-				itemStackIngredients.add(ItemStack.PACKET_CODEC.decode(buf));
+				itemStackIngredients.add(ItemStackIngredient.PACKET_CODEC.decode(buf));
 			}
 			int ingredientsSize = buf.readInt();
 			List<Ingredient> ingredients = new ArrayList<>();
@@ -227,8 +235,8 @@ public class RPGCraftingRecipe implements Recipe<MultipleStackRecipeInput> {
 
 		private static void write(RegistryByteBuf buf, RPGCraftingRecipe recipe) {
 			buf.writeInt(recipe.itemStackIngredients.size());
-			for (ItemStack ingredient : recipe.itemStackIngredients) {
-				ItemStack.PACKET_CODEC.encode(buf, ingredient);
+			for (ItemStackIngredient ingredient : recipe.itemStackIngredients) {
+				ItemStackIngredient.PACKET_CODEC.encode(buf, ingredient);
 			}
 			buf.writeInt(recipe.ingredients.size());
 			for (Ingredient ingredient : recipe.ingredients) {
@@ -243,9 +251,98 @@ public class RPGCraftingRecipe implements Recipe<MultipleStackRecipeInput> {
 		}
 	}
 
+//	public static boolean areItemsAndComponentsEqual(ItemStack stack, ItemStack otherStack) {
+//		if (!stack.isOf(otherStack.getItem())) {
+//			RPGCrafting.info("!stack.isOf(otherStack.getItem())");
+//			return false;
+//		} else {
+//			boolean bl = stack.isEmpty() && otherStack.isEmpty();
+//			RPGCrafting.info("both stacks empty: " + bl);
+//			return bl ? true : Objects.equals(stack.getComponents(), otherStack.getComponents());
+//		}
+//	}
+	public static boolean checkItemStackIngredient(ItemStackIngredient itemStackIngredient, ItemStack itemStack) {
+		if (itemStackIngredient.completeComponentMatch) {
+			RPGCrafting.info("itemStackIngredient.itemStack components: " + itemStackIngredient.itemStack.getComponents());
+			RPGCrafting.info("itemStack components: " + itemStack.getComponents());
+
+			boolean bl = ItemStack.areItemsAndComponentsEqual(itemStackIngredient.itemStack, itemStack);
+			RPGCrafting.info("bl: " + bl);
+			return bl;
+		} else {
+			RPGCrafting.info("!completeComponentMatch");
+			RPGCrafting.info("itemStack.getComponentChanges().entrySet(): " + itemStack.getComponentChanges().entrySet());
+			RPGCrafting.info("itemStackIngredient.itemStack.getComponentChanges().entrySet(): " + itemStackIngredient.itemStack.getComponentChanges().entrySet());
+
+			if (!itemStackIngredient.itemStack.isOf(itemStack.getItem())) {
+				RPGCrafting.info("!stack.isOf(otherStack.getItem())");
+				return false;
+			} else {
+				boolean bl = itemStackIngredient.itemStack.isEmpty() && itemStack.isEmpty();
+				RPGCrafting.info("both stacks empty: " + bl);
+				return bl ? true : itemStack.getComponentChanges().entrySet().containsAll(itemStackIngredient.itemStack.getComponentChanges().entrySet());
+			}
+//				boolean bl =
+//			RPGCrafting.info("bl: " + bl);
+//			return bl;
+//			boolean bl = true;
+//			for (Map.Entry<ComponentType<?>, Optional<?>> entry : componentChanges.entrySet()) {
+//				Optional<?> optional = componentChanges.get(entry.getKey());
+//				RPGCrafting.info("optional: " + optional);
+//				if (optional != null && optional.isPresent()) {
+//					RPGCrafting.info("optional != null && optional.isPresent()");
+//					if (itemStack.contains(entry.getKey())) {
+//						if (optional.get() != itemStack.get(entry.getKey())) {
+//							RPGCrafting.info("optional.get() != itemStack.get(entry.getKey())");
+//							bl = false;
+//							break;
+//						}
+//					} else {
+//						RPGCrafting.info("else");
+//						bl = false;
+//						break;
+//					}
+//				}
+//			}
+//			return bl;
+		}
+	}
+
 	@Override
 	public boolean showNotification() {
 		return this.showNotification;
+	}
+
+	public record ItemStackIngredient(
+			ItemStack itemStack,
+			boolean completeComponentMatch
+	) {
+
+		public static final PacketCodec<RegistryByteBuf, ItemStackIngredient> PACKET_CODEC = new PacketCodec<RegistryByteBuf, ItemStackIngredient>() {
+			public ItemStackIngredient decode(RegistryByteBuf registryByteBuf) {
+				ItemStack itemStack = ItemStack.PACKET_CODEC.decode(registryByteBuf);
+				boolean compareAllComponents = PacketCodecs.BOOL.decode(registryByteBuf);
+				return new ItemStackIngredient(itemStack, compareAllComponents);
+			}
+
+			public void encode(RegistryByteBuf registryByteBuf, ItemStackIngredient itemStackIngredient) {
+				ItemStack.PACKET_CODEC.encode(registryByteBuf, itemStackIngredient.itemStack);
+				PacketCodecs.BOOL.encode(registryByteBuf, itemStackIngredient.completeComponentMatch);
+			}
+		};
+
+		public static final Codec<ItemStackIngredient> CODEC = RecordCodecBuilder.create(instance -> instance.group(
+				ItemStack.VALIDATED_CODEC.fieldOf("itemStack").forGetter(x -> x.itemStack),
+				Codec.BOOL.optionalFieldOf("completeComponentMatch", true).forGetter(x -> x.completeComponentMatch)
+		).apply(instance, ItemStackIngredient::new));
+
+		public ItemStackIngredient(
+				ItemStack itemStack,
+				boolean completeComponentMatch
+		) {
+			this.itemStack = itemStack;
+			this.completeComponentMatch = completeComponentMatch;
+		}
 	}
 
 }