+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+
+- added an optional "completeComponentMatch" boolean field to each entry of the "itemStackIngredients" list, with a default of true
+  - when set to false, an itemStack only has to contain the components listed in the recipe, to be considered a valid ingredient for the recipe (eventual additional components are ignored)
+
 # 1.1.0
 
 - added "itemStackIngredients" to RPGCraftingRecipe. This is a separate list of ingredients that consists of item stacks. When data components are defined for a stack, a stack with the exact same data components is required for the recipe. If no components are defined, any stack of that item type will work. Both types of ingredient lists are active at the same time. Note that existing recipes of the "rpgcrafting:rpg_crafting_recipe" recipe type must be updated!